factoring out test code
